pyproject.toml
[project]
name = "auto-chapter-bar"
version = "0.1.0"
description = "将 SRT 字幕转换为视频章节进度条"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.13"
dependencies = [
    "moviepy>=1.0.3",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "gradio>=4.0.0",
    "pillow>=10.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0.0",
    "loguru>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.14.0",
]

[project.scripts]
acb = "chapterbar.cli:app"
chapterbar = "chapterbar.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chapterbar"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]
ignore = [
    "B008",  # Do not perform function calls in argument defaults (Typer uses this pattern)
]
